Remove NuguInterface namespace
It remove NuguInterface namespace and related directories.

The header and source files which included in NuguInterface
are moved to NuguClientKit namespace.

The CapabilityCreator is divided to CapabilityFactory and
NuguCoreContainer as following each namespace.
