+++ src/capability/tts_agent.cc
@@ -18,13 +18,13 @@
 
 #include "base/nugu_log.h"
 #include "base/nugu_uuid.h"
-#include "interface/nugu_configuration.hh"
+#include "clientkit/nugu_configuration.hh"
 
 #include "tts_agent.hh"
 
 namespace NuguCapability {
 
-using namespace NuguInterface;
+using namespace NuguClientKit;
 
 static const char* CAPABILITY_NAME = "TTS";
 static const char* CAPABILITY_VERSION = "1.0";
