+++ src/core/nugu_core_container.cc
@@ -1,0 +1,34 @@
+/*
+ * Copyright (c) 2019 SK Telecom Co., Ltd. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+#include "network_manager.hh"
+#include "wakeup_handler.hh"
+
+#include "nugu_core_container.hh"
+
+namespace NuguCore {
+
+IWakeupHandler* NuguCoreContainer::createWakeupHandler()
+{
+    return new WakeupHandler();
+}
+
+INetworkManager* NuguCoreContainer::createNetworkManager()
+{
+    return new NetworkManager();
+}
+
+} // NuguCore
