+++ src/core/wakeup_detector.cc
@@ -17,13 +17,13 @@
 #include <keyword_detector.h>
 
 #include "base/nugu_log.h"
-#include "interface/nugu_configuration.hh"
+#include "clientkit/nugu_configuration.hh"
 
 #include "wakeup_detector.hh"
 
 namespace NuguCore {
 
-using namespace NuguInterface;
+using namespace NuguClientKit;
 
 WakeupDetector::WakeupDetector()
 {
