--- src/interface/CMakeLists.txt
@@ -1,2 +1,0 @@
-FILE(GLOB_RECURSE INTERFACE_SRC *.cc)
-ADD_LIBRARY(objinterface OBJECT ${INTERFACE_SRC})
